Switch back to WebSocket api, but keep custom http
Done to make the web simulator simpler. The plan had been to use the
HTTP + SSE api even in the simulator by proxying it through a service
worker that would `postMessage` back to the page running the simulator.
This was possible, but pretty janky and the configurator had to be built
differently for the simulator anyway, since it needs a different base
path. Instead, the configurator can simply `postMessage` back to the
simulator page directly, and there is less reason to stick with the HTTP
+ SSE api, so back we go.